+++ src/app/component/admin/admin.menus.ts
@@ -1,13 +1,7 @@
 import { sitesCategory } from "@component/sites/sites.menus";
 import { Category, MenuLink, MenuRoute } from "@interfaces/menusInterfaces";
 import { StrongRoute } from "@interfaces/strongRoute";
-import {
-  defaultAudioIcon,
-  defaultDeleteIcon,
-  defaultEditIcon,
-  defaultNewIcon,
-  isAdminPredicate,
-} from "src/app/app.menus";
+import { defaultAudioIcon, isAdminPredicate } from "src/app/app.menus";
 
 export const adminRoute = StrongRoute.Base.add("admin");
 export const adminCategory: Category = {
@@ -39,56 +33,6 @@ export const adminOrphanSitesMenuItem = MenuRoute({
   route: adminRoute.add("sites"),
   tooltip: () => "Manage orphaned sites",
   parent: adminDashboardMenuItem,
-  predicate: isAdminPredicate,
-});
-
-/**
- * Admin Tags
- */
-
-const adminTagsRoute = adminRoute.add("tags");
-
-export const adminTagsCategory: Category = {
-  icon: ["fas", "tag"],
-  label: "Tags",
-  route: adminTagsRoute,
-};
-
-export const adminTagsMenuItem = MenuRoute({
-  icon: ["fas", "tag"],
-  label: "Tags",
-  route: adminTagsRoute,
-  tooltip: () => "Manage tags",
-  parent: adminDashboardMenuItem,
-  predicate: isAdminPredicate,
-});
-
-export const adminNewTagMenuItem = MenuRoute({
-  icon: defaultNewIcon,
-  label: "New Tag",
-  route: adminTagsRoute.add("new"),
-  tooltip: () => "Create a new tag",
-  parent: adminTagsMenuItem,
-  predicate: isAdminPredicate,
-});
-
-const adminTagRoute = adminTagsRoute.add(":tagId");
-
-export const adminEditTagMenuItem = MenuRoute({
-  icon: defaultEditIcon,
-  label: "Edit Tag",
-  route: adminTagRoute.add("edit"),
-  tooltip: () => "Edit an existing tag",
-  parent: adminTagsMenuItem,
-  predicate: isAdminPredicate,
-});
-
-export const adminDeleteTagMenuItem = MenuRoute({
-  icon: defaultDeleteIcon,
-  label: "Delete Tag",
-  route: adminTagRoute.add("delete"),
-  tooltip: () => "Delete an existing tag",
-  parent: adminTagsMenuItem,
   predicate: isAdminPredicate,
 });
