+++ src/app/component/admin/tags/delete/delete.component.spec.ts
@@ -3,7 +3,6 @@ import { ActivatedRoute, Router } from "@angular/router";
 import { RouterTestingModule } from "@angular/router/testing";
 import { ApiErrorDetails } from "@baw-api/api.interceptor.service";
 import { tagResolvers, TagsService } from "@baw-api/tags.service";
-import { adminTagsMenuItem } from "@component/admin/admin.menus";
 import { Tag } from "@models/Tag";
 import { SharedModule } from "@shared/shared.module";
 import { ToastrService } from "ngx-toastr";
@@ -11,6 +10,7 @@ import { BehaviorSubject } from "rxjs";
 import { appLibraryImports } from "src/app/app.module";
 import { mockActivatedRoute, testBawServices } from "src/app/test.helper";
 import { assertFormErrorHandling } from "src/testHelpers";
+import { adminTagsMenuItem } from "../tags.menus";
 import { AdminTagsDeleteComponent } from "./delete.component";
 
 describe("AdminTagsDeleteComponent", () => {
