+++ src/app/component/admin/tags/edit/edit.component.ts
@@ -2,12 +2,6 @@ import { Component, OnInit } from "@angular/core";
 import { ActivatedRoute, Router } from "@angular/router";
 import { tagResolvers, TagsService, TagType } from "@baw-api/tags.service";
 import {
-  adminDeleteTagMenuItem,
-  adminEditTagMenuItem,
-  adminTagsCategory,
-  adminTagsMenuItem,
-} from "@component/admin/admin.menus";
-import {
   defaultSuccessMsg,
   FormTemplate,
 } from "@helpers/formTemplate/formTemplate";
@@ -17,6 +11,12 @@ import { List } from "immutable";
 import { ToastrService } from "ngx-toastr";
 import { adminTagsMenuItemActions } from "../list/list.component";
 import { fields } from "../tag.json";
+import {
+  adminDeleteTagMenuItem,
+  adminEditTagMenuItem,
+  adminTagsCategory,
+  adminTagsMenuItem,
+} from "../tags.menus";
 
 const tagKey = "tag";
 const tagTypesKey = "tagTypes";
